backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.2.0
openpyxl==3.1.5
python-dateutil==2.9.0
msal
httpx
python-dotenv
